+++ Knowledge_Plugin/DOKE/config/ml1m/popneg_empty.json
@@ -1,7 +1,7 @@
 {
-    "sequential_data_path": "../../data/ml1m/sequential_data.txt",
-    "candidate_data_path": "../../data/ml1m/negative_samples_pop.txt",
-    "meta_data_path": "../../data/ml1m/metadata.json",
+    "sequential_data_path": "../data/ml1m/sequential_data.txt",
+    "candidate_data_path": "../data/ml1m/negative_samples_pop.txt",
+    "meta_data_path": "../data/ml1m/metadata.json",
     "topk": 200,
     "max_his_len": 50,
     "template": [
