+++ Knowledge_Plugin/DOKE/config/ml1m/popneg_global_I2I.json
@@ -1,8 +1,8 @@
 {
-    "sequential_data_path": "../../data/ml1m/sequential_data.txt",
-    "candidate_data_path": "../../data/ml1m/negative_samples_pop.txt",
-    "meta_data_path": "../../data/ml1m/metadata.json",
-    "global_cf_data_path": "data/ml1m/global_CF.json",
+    "sequential_data_path": "../data/ml1m/sequential_data.txt",
+    "candidate_data_path": "../data/ml1m/negative_samples_pop.txt",
+    "meta_data_path": "../data/ml1m/metadata.json",
+    "global_cf_data_path": "../data/ml1m/global_CF.json",
     "topk": 200,
     "max_his_len": 50,
     "template": [
