+++ Knowledge_Plugin/DOKE/config/ml1m/popneg_his-can_I2I.json
@@ -1,8 +1,8 @@
 {
-    "sequential_data_path": "../../data/ml1m/sequential_data.txt",
-    "candidate_data_path": "../../data/ml1m/negative_samples_pop.txt",
-    "meta_data_path": "../../data/ml1m/metadata.json",
-    "cf_data_path": "data/ml1m/MF_CF_candidate_pop.json",
+    "sequential_data_path": "../data/ml1m/sequential_data.txt",
+    "candidate_data_path": "../data/ml1m/negative_samples_pop.txt",
+    "meta_data_path": "../data/ml1m/metadata.json",
+    "cf_data_path": "../data/ml1m/MF_CF_candidate_pop.json",
     "topk": 200,
     "max_his_len": 50,
     "template": [
