+++ Knowledge_Plugin/DOKE/config/ml1m/popneg_his_I2I.json
@@ -2,7 +2,7 @@
     "sequential_data_path": "../data/ml1m/sequential_data.txt",
     "candidate_data_path": "../data/ml1m/negative_samples_pop.txt",
     "meta_data_path": "../data/ml1m/metadata.json",
-    "cf_data_path": "data/ml1m/normalized_CF.json",
+    "cf_data_path": "../data/ml1m/normalized_CF.json",
     "topk": 200,
     "max_his_len": 50,
     "template": [
