+++ Knowledge_Plugin/preprocess/step2-Base_models/SASRec/logs/ml1m.log
@@ -1,5 +1,5 @@
 average sequence length: 93.34
-Evaluating and Testing...
+10 Evaluating and Testing...
 testing...
-eval result: ndcg@1: 0.44, ndcg@5: 0.6297, ndcg@10: 0.6755, ndcg@20: 0.6959, ndcg@50: 0.6959, ndcg@100: 0.6959, hit@1: 0.44, hit@5: 0.78, hit@10: 0.92, hit@20: 1.0, hit@50: 1.0, hit@100: 1.0, precision@1: 0.44, precision@5: 0.5789, precision@10: 0.5982, precision@20: 0.6039, precision@50: 0.6039, precision@100: 0.6039, recall@1: 0.44, recall@5: 0.78, recall@10: 0.92, recall@20: 1.0, recall@50: 1.0, recall@100: 1.0, 
+eval result: ndcg@1: 0.242, ndcg@5: 0.4534, ndcg@10: 0.5205, ndcg@20: 0.5563, ndcg@50: 0.5563, ndcg@100: 0.5563, hit@1: 0.242, hit@5: 0.652, hit@10: 0.859, hit@20: 1.0, hit@50: 1.0, hit@100: 1.0, precision@1: 0.242, precision@5: 0.388, precision@10: 0.4157, precision@20: 0.4256, precision@50: 0.4256, precision@100: 0.4256, recall@1: 0.242, recall@5: 0.652, recall@10: 0.859, recall@20: 1.0, recall@50: 1.0, recall@100: 1.0, 
 Done
